// generated/models/ApiArtworkDocumentationPublicPreview.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiArtworkDocumentationAssetLink } from '../models/ApiArtworkDocumentationAssetLink';
import { ApiArtworkDocumentationPreviewModule } from '../models/ApiArtworkDocumentationPreviewModule';
import { HttpFile } from '../http/http';

export class ApiArtworkDocumentationPublicPreview {
    'context_id': string;
    'draft_version': number;
    'modules': { [key: string]: ApiArtworkDocumentationPreviewModule; };
    'asset_links': Array<ApiArtworkDocumentationAssetLink>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "context_id",
            "baseName": "context_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "draft_version",
            "baseName": "draft_version",
            "type": "number",
            "format": ""
        },
        {
            "name": "modules",
            "baseName": "modules",
            "type": "{ [key: string]: ApiArtworkDocumentationPreviewModule; }",
            "format": ""
        },
        {
            "name": "asset_links",
            "baseName": "asset_links",
            "type": "Array<ApiArtworkDocumentationAssetLink>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiArtworkDocumentationPublicPreview.attributeTypeMap;
    }

    public constructor() {
    }
}